standalone program to process root files produced with the analysis module in this package.
The user has to use cmsRun and EcalHVScanAnalysis module to produce a root file
that contains a TTree with the relevant information.

this standalone program is then used to quickly produce histograms and plots from the TTree
w/o running again on the raw data which takes relatively a long time.