Start of reorganization/cleanup:
- Move most globals (most importantly those dealing with the
parameterizations) from AsymFunctions.C/h inside the AsymFitManager so
that there's less trouble including AsymFunctions.h in different
places.

- Piece out the parametrization code into its own module so that it
can be easily run in a separate step. This is necessary if running on
MC that is not on local disk, but also makes the code and workflow
nicer.

- Remove/clean up a lot of MC-specific config parameters, and let
AsymFitManager handle most of what remains.