Old zp2mu framework removal: Zprime2muMassReach moved into plugins/Zprime2muMassReach.cc.
Also, instead of bringing Zprime2muMassReach into the new framework,
it is easier to keep it as-is by having the (needed bits of the)
Zprime2muAnalysis class defined in Zprime2muMassReach.cc. The
Zprime2muAnalysis class is not left in its own file so that no one
thinks it is available to be or should be used.