write event weight for artificial tau events
To account correctly for the number of attempts until an tautau event where the decay products exceed a certain visible transverse momenta was created, many toy events are generated. The fraction of events passing the vis_pT cut is written as a double to the event, the first event fulfilling the requirement is used for further processing.