Create separate product for a plugin.
It looks like some of the utilities in Calibration/Tools depend on stuff in
this package (which is a plugin). This is not actually allowed by the CMS
policy (and actively disable). However in this case it works because the
shared code is inlined when compiling -O2 and link time checks fail.
However when compiling with -O0 the code is not inlined anymore and CMSSW
fails to build Calibration/Tools for this reason.

This fixes the problem by creating a separate product for the plugin and
making the DSO which contains the common code a normal library.